refactor: rename command to runner
